// sanity/sanity.types.ts
export type ACTIVE_DAILY_PROMPTS_QUERYResult = Array<{
  _id: string
  title: string | null
  prompt: string | null
  emoji: string | null
  category: {
    title: string | null
    color: string | null
  } | null
  suggestedMood: 'happy' | 'neutral' | 'sad' | 'very-happy' | 'very-sad' | null
  isActive: boolean | null
  weight: number | null
  tags: Array<string> | null
  createdAt: string | null
}>